whitepaper
In times of war, ninja will come together to defend their villages and land. The clearest characteristic of a ninja is their ability to manipulate chakra to create and use techniques. Naruto  tells the story of Naruto Uzumaki, a young ninja who seeks recognition from his peers and dreams of becoming the Hokage, the leader of his village. NINJAIN Inu is a community driven platform promoting financial well-being by empowering people of all ages and cultures to be in control of their money and create their own wealth opportunities.